requirements.txt
appium-python-client==5.2.4
selenium==4.26.0
pytest==9.0.2